+++ src/ladb_opencutlist/ruby/worker/materials/materials_update_worker.rb
@@ -16,6 +16,7 @@ module Ladb::OpenCutList
       @texture_height = material_data.fetch('texture_height')
 
       attributes = material_data.fetch('attributes')
+      @description = attributes.fetch('description', '')
       @type = MaterialAttributes.valid_type(attributes.fetch('type'))
       @thickness = attributes.fetch('thickness')
       @length_increase = attributes.fetch('length_increase')
@@ -111,6 +112,7 @@ module Ladb::OpenCutList
 
       # Update attributes
       material_attributes = MaterialAttributes.new(material)
+      material_attributes.description = @description
       material_attributes.type = @type
       material_attributes.thickness = @thickness
       material_attributes.length_increase = @length_increase